control/e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "robotframework>=7.4.1",
    "robotframework-imaplibrary2>=0.4.11",
    "robotframework-requests>=0.9.7",
    "robotframework-robocop>=7.2.0",
]
